random test + more consistent rejection

+++ tests/assembly/assembly.rs
@@ -40,21 +40,23 @@ const BOOT_DURATION: Duration = Duration::from_secs(10);
 
 macro_rules! start_server {
     () => {{
-        let output = build_cmd("typedb-extracted/typedb server --development-mode.enabled=true")
+        let process = build_cmd("typedb-extracted/typedb server --development-mode.enabled=true")
+            .stdout(std::process::Stdio::piped())
+            .stderr(std::process::Stdio::piped())
             .spawn()
             .expect("Failed to run console script");
         std::thread::sleep(BOOT_DURATION);
-        output
+        process
     }};
     ($env:expr => $value:expr) => {{
-        let output = build_cmd("typedb-extracted/typedb server --development-mode.enabled=true")
+        let process = build_cmd("typedb-extracted/typedb server --development-mode.enabled=true")
             .env($env, $value)
             .stdout(std::process::Stdio::null())
-            .stdout(std::process::Stdio::null())
+            .stderr(std::process::Stdio::null())
             .spawn()
             .expect("Failed to run console script");
         std::thread::sleep(BOOT_DURATION);
-        output
+        process
     }};
 }
 
@@ -81,13 +83,16 @@ fn test_assembly() {
 #[test]
 fn test_fail_point_always() {
     for fail_point in fail_point::ALL {
-        extract_typedb();
         let directive = format!("{fail_point}=panic");
-        let mut server_process = start_server!(fail_point::FAIL_POINT_ENV => &directive);
-        run_test_against_server(&mut server_process);
 
+        extract_typedb();
+        let mut server_process = start_server!(fail_point::FAIL_POINT_ENV => &directive);
+        setup();
+
+        run_test_against_server(&mut server_process);
         if server_process.try_wait().unwrap().is_none() {
             kill_process(server_process).unwrap();
+            // some fail points are only triggered on second boot
             let mut server_process = start_server!(fail_point::FAIL_POINT_ENV => directive);
             if server_process.try_wait().unwrap().is_none() {
                 kill_process(server_process).unwrap();
@@ -95,14 +100,51 @@ fn test_fail_point_always() {
             }
         }
 
-        let mut server_process = start_server!();
-        match server_process.try_wait().unwrap() {
-            None => _ = kill_process(server_process).unwrap(),
-            Some(_) => {
-                let server_process_output = kill_process(server_process).unwrap();
-                eprintln!("Server process output:\n{:?}", server_process_output);
-                panic!("Server could not boot after fail point {fail_point} triggered");
+        assert_boots(fail_point);
+    }
+}
+
+#[test]
+fn test_fail_point_chance() {
+    for fail_point in fail_point::ALL {
+        let directive = format!("{fail_point}=90%5*print->panic"); // 10% chance to panic, but guaranteed on the 6th
+
+        extract_typedb();
+        let mut server_process = start_server!(fail_point::FAIL_POINT_ENV => &directive);
+        setup();
+
+        for _ in 0..10 {
+            if server_process.try_wait().unwrap().is_some() {
+                break; // crashed
             }
+
+            kill_process(server_process).unwrap();
+            server_process = start_server!(fail_point::FAIL_POINT_ENV => &directive);
+
+            for _ in 0..6 {
+                run_test_against_server(&mut server_process);
+                if server_process.try_wait().unwrap().is_some() {
+                    break; // crashed
+                }
+            }
+        }
+
+        if server_process.try_wait().unwrap().is_none() {
+            kill_process(server_process).unwrap(); // did not crash; must be a fail point that triggers at most once per run and we got unlucky
+        }
+
+        assert_boots(fail_point);
+    }
+}
+
+fn assert_boots(fail_point: &str) {
+    let mut server_process = start_server!();
+    match server_process.try_wait().unwrap() {
+        None => _ = kill_process(server_process).unwrap(),
+        Some(_) => {
+            let server_process_output = kill_process(server_process).unwrap();
+            eprintln!("Server process output:\n{:?}", server_process_output);
+            panic!("Server could not boot after fail point {fail_point} triggered");
         }
     }
 }
@@ -128,6 +170,16 @@ fn extract_typedb() {
     }
 }
 
+fn setup() {
+    let command = "
+        database create foo
+        transaction schema foo
+        define entity person, owns name @card(0..1); attribute name, value string; end;
+        commit
+        ";
+    build_console_command(command).output().expect("Failed to run console script");
+}
+
 fn run_test_against_server(server_process: &mut Child) {
     enum Instruction {
         Command(&'static str),
@@ -136,35 +188,40 @@ fn run_test_against_server(server_process: &mut Child) {
     }
 
     let instructions = [
-        Instruction::Command("database create foo"),
         Instruction::Command(
-            r#"
-transaction schema foo
-define entity person, owns name @card(0..1); attribute name, value string; end;
-commit"#,
+            "
+            transaction schema foo
+            define entity person, owns name @card(0..1); attribute name, value string; end;
+            commit
+            ",
         ),
         Instruction::Sleep(CHECKPOINT_INTERVAL),
         Instruction::Command(
             r#"
-transaction write foo
-insert $john isa person, has name "John"; end;
-commit"#,
+            transaction write foo
+            insert $john isa person, has name "John"; end;
+            commit
+            "#,
         ),
         Instruction::Sleep(CHECKPOINT_INTERVAL),
         Instruction::Parallel(
-            r#"
-transaction write foo
-match $john isa person, has name $name;
-delete has $name of $john; end;
-commit"#,
+            "
+            transaction write foo
+            match $john isa person, has name $name;
+            delete has $name of $john;
+            insert $john has $name;
+            end;
+            commit
+            ",
         ),
         Instruction::Command("database delete foo"),
         Instruction::Command("database create foo"),
         Instruction::Command(
-            r#"
-transaction schema foo
-define entity person, owns name @card(0..1); attribute name, value string; end;
-commit"#,
+            "
+            transaction schema foo
+            define entity person, owns name @card(0..1); attribute name, value string; end;
+            commit
+            ",
         ),
     ];
 
@@ -177,10 +234,12 @@ commit"#,
                 }
             }
             Instruction::Parallel(command) => {
-                let mut cmd = build_console_command(command);
-                let mut cmd2 = build_console_command(command);
-                thread::spawn(move || cmd.output().unwrap());
-                cmd2.output().unwrap();
+                let cmds = std::array::from_fn::<_, 10, _>(|_| build_console_command(command));
+                let threads =
+                    cmds.into_iter().map(|mut cmd| thread::spawn(move || cmd.output().unwrap())).collect::<Vec<_>>();
+                for thread in threads {
+                    thread.join().unwrap();
+                }
             }
             Instruction::Sleep(duration) => wait_process_timeout(server_process, duration).unwrap(),
         }
